Revert "Refactor core metrics into backend modules"
