+++ backend/workout_session.py
@@ -2,8 +2,13 @@ import sqlite3
 import time
 from pathlib import Path
 
-from . import DEFAULT_SETS_PER_EXERCISE, DEFAULT_REST_DURATION, DEFAULT_DB_PATH
-from .metrics import get_metrics_for_exercise, get_metrics_for_preset
+from core import (
+    get_metrics_for_exercise,
+    get_metrics_for_preset,
+    DEFAULT_SETS_PER_EXERCISE,
+    DEFAULT_REST_DURATION,
+    DEFAULT_DB_PATH,
+)
 
 
 class WorkoutSession:
