+++ tests/test_database_helpers.py
@@ -1,4 +1,5 @@
-from backend import exercises, metrics, presets
+import core
+from backend import exercises
 from backend.exercise import Exercise
 
 
@@ -23,20 +24,20 @@ def test_get_all_exercises(sample_db):
 
 
 def test_get_metrics_for_exercise(sample_db):
-    default = metrics.get_metrics_for_exercise("Bench Press", db_path=sample_db)
-    override = metrics.get_metrics_for_exercise("Bench Press", preset_name="Push Day", db_path=sample_db)
+    default = core.get_metrics_for_exercise("Bench Press", db_path=sample_db)
+    override = core.get_metrics_for_exercise("Bench Press", preset_name="Push Day", db_path=sample_db)
 
     def get_timing(metrics, name):
         return next(m for m in metrics if m["name"] == name)["input_timing"]
 
     assert get_timing(default, "Reps") == "post_set"
     assert get_timing(override, "Reps") == "pre_set"
-    assert metrics.get_metrics_for_exercise("Unknown", db_path=sample_db) == []
+    assert core.get_metrics_for_exercise("Unknown", db_path=sample_db) == []
 
 
 def test_load_workout_presets(sample_db):
-    presets_list = presets.load_workout_presets(sample_db)
-    assert presets_list == [
+    presets = core.load_workout_presets(sample_db)
+    assert presets == [
         {
             "name": "Push Day",
             "exercises": [
