+++ tests/test_exercise_model.py
@@ -1,3 +1,4 @@
+import core
 from backend import exercises
 from backend.exercise import Exercise
