+++ tests/test_workout_db.py
@@ -4,7 +4,8 @@ import sys
 import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-from backend import exercises, metrics, presets
+import core
+from backend import exercises
 
 
 def create_empty_db(path: Path) -> None:
@@ -87,7 +88,7 @@ def sample_db(tmp_db: Path) -> Path:
 
 
 def test_get_metric_type_schema(tmp_db: Path):
-    fields = metrics.get_metric_type_schema(db_path=tmp_db)
+    fields = core.get_metric_type_schema(db_path=tmp_db)
     names = {f["name"] for f in fields}
     assert names == {"name", "type", "input_timing", "is_required", "scope", "description", "enum_values_json"}
     type_opts = next(f["options"] for f in fields if f["name"] == "type")
@@ -109,8 +110,8 @@ def test_get_all_exercises_and_details(sample_db: Path):
 
 
 def test_load_workout_presets(sample_db: Path):
-    presets_list = presets.load_workout_presets(db_path=sample_db)
-    assert presets_list == [
+    presets = core.load_workout_presets(db_path=sample_db)
+    assert presets == [
         {
             "name": "Push Day",
             "exercises": [
@@ -122,7 +123,7 @@ def test_load_workout_presets(sample_db: Path):
 
 
 def test_add_and_remove_metric_from_exercise(sample_db: Path):
-    metrics.add_metric_to_exercise("Bench Press", "Reps", db_path=sample_db)
+    core.add_metric_to_exercise("Bench Press", "Reps", db_path=sample_db)
     conn = sqlite3.connect(sample_db)
     cur = conn.cursor()
     cur.execute(
@@ -130,7 +131,7 @@ def test_add_and_remove_metric_from_exercise(sample_db: Path):
     )
     count = cur.fetchone()[0]
     assert count == 1
-    metrics.remove_metric_from_exercise("Bench Press", "Reps", db_path=sample_db)
+    core.remove_metric_from_exercise("Bench Press", "Reps", db_path=sample_db)
     cur.execute(
         """SELECT COUNT(*) FROM library_exercise_metrics em JOIN library_exercises e ON em.exercise_id = e.id JOIN library_metric_types mt ON em.metric_type_id = mt.id WHERE e.name='Bench Press' AND mt.name='Reps' AND em.deleted = 0 AND e.deleted = 0 AND mt.deleted = 0"""
     )
