+++ ui/popups.py
@@ -19,7 +19,8 @@ import string
 import re
 import sqlite3
 
-from backend import metrics as metrics_lib, presets as presets_lib, DEFAULT_DB_PATH
+import core
+from core import DEFAULT_DB_PATH
 
 # Order of fields for metric editing popups
 METRIC_FIELD_ORDER = [
@@ -74,7 +75,7 @@ class AddMetricPopup(MDDialog):
     # Building widgets for both modes
     # ------------------------------------------------------------------
     def _build_select_widgets(self):
-        metrics = metrics_lib.get_all_metric_types()
+        metrics = core.get_all_metric_types()
         existing = {m.get("name") for m in self.screen.exercise_obj.metrics}
         metrics = [
             m
@@ -101,7 +102,7 @@ class AddMetricPopup(MDDialog):
         default_height = dp(48)
         self.input_widgets = {}
 
-        schema = metrics_lib.get_metric_type_schema()
+        schema = core.get_metric_type_schema()
         if not schema:
             schema = [
                 {"name": "name"},
@@ -247,7 +248,7 @@ class AddMetricPopup(MDDialog):
         popup.open()
 
     def add_metric(self, name, *args):
-        metric_defs = metrics_lib.get_all_metric_types()
+        metric_defs = core.get_all_metric_types()
         for m in metric_defs:
             if m["name"] == name:
                 self.screen.exercise_obj.add_metric(m)
@@ -306,7 +307,7 @@ class AddMetricPopup(MDDialog):
 
         db_path = DEFAULT_DB_PATH
         try:
-            metrics_lib.add_metric_type(
+            core.add_metric_type(
                 metric["name"],
                 metric["type"],
                 metric["input_timing"],
@@ -365,7 +366,7 @@ class EditMetricPopup(MDDialog):
         default_height = dp(48)
         self.input_widgets = {}
 
-        schema = metrics_lib.get_metric_type_schema()
+        schema = core.get_metric_type_schema()
         if not schema:
             schema = [
                 {"name": "name"},
@@ -607,7 +608,7 @@ class EditMetricPopup(MDDialog):
             def on_save(*a):
                 metric_saved = self.screen.exercise_obj.had_metric(self.metric["name"])
                 if cb_type.active:
-                    metrics_lib.update_metric_type(
+                    core.update_metric_type(
                         self.metric["name"],
                         mtype=updates.get("type"),
                         input_timing=updates.get("input_timing"),
@@ -618,7 +619,7 @@ class EditMetricPopup(MDDialog):
                         db_path=db_path,
                     )
                     if metric_saved:
-                        metrics_lib.set_exercise_metric_override(
+                        core.set_exercise_metric_override(
                             self.screen.exercise_obj.name,
                             self.metric["name"],
                             is_user_created=self.screen.exercise_obj.is_user_created,
@@ -626,7 +627,7 @@ class EditMetricPopup(MDDialog):
                         )
                 elif cb_ex.active:
                     if metric_saved:
-                        metrics_lib.set_exercise_metric_override(
+                        core.set_exercise_metric_override(
                             self.screen.exercise_obj.name,
                             self.metric["name"],
                             mtype=updates.get("type"),
@@ -639,7 +640,7 @@ class EditMetricPopup(MDDialog):
                         )
                 else:
                     preset_name = app.preset_editor.preset_name if app else ""
-                    metrics_lib.set_section_exercise_metric_override(
+                    core.set_section_exercise_metric_override(
                         preset_name,
                         self.screen.section_index,
                         self.screen.exercise_obj.name,
@@ -680,9 +681,9 @@ class EditMetricPopup(MDDialog):
 
             rows = []
             cb_default = None
-            if metrics_lib.is_metric_type_user_created(
+            if core.is_metric_type_user_created(
                 self.metric["name"], db_path=db_path
-            ) and metrics_lib.uses_default_metric(
+            ) and core.uses_default_metric(
                 self.screen.exercise_obj.name,
                 self.metric["name"],
                 is_user_created=self.screen.exercise_obj.is_user_created,
@@ -706,7 +707,7 @@ class EditMetricPopup(MDDialog):
                 )
                 rows.append(row)
 
-            presets = presets_lib.find_presets_using_exercise(
+            presets = core.find_presets_using_exercise(
                 self.screen.exercise_obj.name, db_path=db_path
             )
             cb_presets = None
@@ -741,7 +742,7 @@ class EditMetricPopup(MDDialog):
                         self.metric["name"]
                     )
                     if cb_default and cb_default.active:
-                        metrics_lib.update_metric_type(
+                        core.update_metric_type(
                             self.metric["name"],
                             mtype=updates.get("type"),
                             input_timing=updates.get("input_timing"),
@@ -752,7 +753,7 @@ class EditMetricPopup(MDDialog):
                             db_path=db_path,
                         )
                         if metric_saved:
-                            metrics_lib.set_exercise_metric_override(
+                            core.set_exercise_metric_override(
                                 self.screen.exercise_obj.name,
                                 self.metric["name"],
                                 is_user_created=self.screen.exercise_obj.is_user_created,
@@ -760,7 +761,7 @@ class EditMetricPopup(MDDialog):
                             )
                     else:
                         if metric_saved:
-                            metrics_lib.set_exercise_metric_override(
+                            core.set_exercise_metric_override(
                                 self.screen.exercise_obj.name,
                                 self.metric["name"],
                                 mtype=updates.get("type"),
@@ -772,7 +773,7 @@ class EditMetricPopup(MDDialog):
                                 db_path=db_path,
                             )
                     if cb_presets and cb_presets.active:
-                        presets_lib.apply_exercise_changes_to_presets(
+                        core.apply_exercise_changes_to_presets(
                             self.screen.exercise_obj,
                             presets,
                             db_path=db_path,
@@ -801,7 +802,7 @@ class EditMetricPopup(MDDialog):
             else:
                 apply_updates()
 
-        elif metrics_lib.is_metric_type_user_created(self.metric["name"], db_path=db_path):
+        elif core.is_metric_type_user_created(self.metric["name"], db_path=db_path):
             dialog = None
 
             def cancel_action(*a):
@@ -821,7 +822,7 @@ class EditMetricPopup(MDDialog):
             def on_save(*a):
                 metric_saved = self.screen.exercise_obj.had_metric(self.metric["name"])
                 if checkbox.active:
-                    metrics_lib.update_metric_type(
+                    core.update_metric_type(
                         self.metric["name"],
                         mtype=updates.get("type"),
                         input_timing=updates.get("input_timing"),
@@ -832,7 +833,7 @@ class EditMetricPopup(MDDialog):
                         db_path=db_path,
                     )
                     if metric_saved:
-                        metrics_lib.set_exercise_metric_override(
+                        core.set_exercise_metric_override(
                             self.screen.exercise_obj.name,
                             self.metric["name"],
                             is_user_created=self.screen.exercise_obj.is_user_created,
@@ -840,7 +841,7 @@ class EditMetricPopup(MDDialog):
                         )
                 else:
                     if metric_saved:
-                        metrics_lib.set_exercise_metric_override(
+                        core.set_exercise_metric_override(
                             self.screen.exercise_obj.name,
                             self.metric["name"],
                             mtype=updates.get("type"),
