+++ ui/screens/exercise_library.py
@@ -18,7 +18,9 @@ from kivymd.uix.dialog import MDDialog
 from kivymd.uix.button import MDRaisedButton
 
 import os
-from backend import exercises, metrics as metrics_lib, DEFAULT_DB_PATH
+import core
+from backend import exercises
+from core import DEFAULT_DB_PATH
 
 
 class ExerciseLibraryScreen(MDScreen):
@@ -64,7 +66,7 @@ class ExerciseLibraryScreen(MDScreen):
             and self.metric_cache_version != getattr(app, "metric_library_version", 0)
         ):
             db_path = DEFAULT_DB_PATH
-            self.all_metrics = metrics_lib.get_all_metric_types(
+            self.all_metrics = core.get_all_metric_types(
                 db_path, include_user_created=True
             )
             if app:
@@ -147,7 +149,7 @@ class ExerciseLibraryScreen(MDScreen):
             and self.metric_cache_version != getattr(app, "metric_library_version", 0)
         ):
             db_path = DEFAULT_DB_PATH
-            self.all_metrics = metrics_lib.get_all_metric_types(
+            self.all_metrics = core.get_all_metric_types(
                 db_path, include_user_created=True
             )
             if app:
@@ -283,7 +285,7 @@ class ExerciseLibraryScreen(MDScreen):
         def do_delete(*args):
             db_path = DEFAULT_DB_PATH
             try:
-                metrics_lib.delete_metric_type(
+                core.delete_metric_type(
                     metric_name, db_path=db_path, is_user_created=True
                 )
                 app = MDApp.get_running_app()
